Reworking PyMsBayes environment and scripts.
Reworking how the Python 2 environment is created and used.  Using
virtualenvwrapper (rather than virtualenv directly) to setup up the
Python environment. This difference required updating of all the shell
scripts for running pymsbayes.

For some reason the calculation of the summary stats fails when
pymsbayes is installed in an environment created with virtualenv (they
are all zero). Using virtualenvwrapper fixes this problem.