Adding pymsbayes info files.
Adding info files output during pymsbayes analsyes. These document the
conditions of the analyses and the run durations. I don't think it is
worth committing any of the other outputs of the pymsbayes analyses do
to the enormous disk space they take up. These can always be
re-generated with the python environment and scripts provided in the
repo.