Adding `random` as required package.
When attempting to run the script `fastas2nex` from pyenv/bin, it was
killed because the random package wasn't loaded in the
fastas2alignment.py. I am guessing that by adding random as a package to
load in this file future users who run the setup bash script will get
a python environment with the random package.