Adding TSV files of empirical dpp-msbayes results.
These files contain posterior and prior probs and bayes factors for the
number of divergence events to be plotted with pyco-sumevents to compare
to the ecoevolity results.

The data in these files for the Philippines analyses come directly from
the dpp-msbayes results that went into the dpp-msbayes paper.  The data
for the Mediterranean results were approximated by measuring the
posterior probability bar heights in Figure 5 of the Papadopoulou and
Knowles paper using Inkscape (the prior probabilities are known from the
Dirichlet Process). Also adding the inkscape file.